+++ AtraCore/Framework/Caches/NPCCache.cs
@@ -78,5 +78,8 @@ public static class NPCCache
         return npc;
     }
 
+    /// <summary>
+    /// Clears the cache.
+    /// </summary>
     internal static void Reset() => cache.Clear();
 }