+++ FixPigRandom/ModEntry.cs
@@ -78,8 +78,7 @@ internal sealed class ModEntry : Mod
         if (!Cache.TryGetValue(id, out Random? random))
         {
             modMonitor.DebugOnlyLog($"Cache miss: {id}", LogLevel.Info);
-            random = RandomUtils.GetSeededRandom(2, (int)(id >> 1));
-            Cache[id] = random;
+            Cache[id] = random = RandomUtils.GetSeededRandom(2, (int)(id >> 1));
         }
 
         return random;