+++ HighlightEmptyMachines/HighlightEmptyMachines/docs/Changelog.md
@@ -3,6 +3,9 @@
 
 TODO: optionally highlight **buildings** that have empty machines within them?
 
+### Version 0.2.2
+* Recompiled for new AtraCore, my apologies!
+
 ### Version 0.2.1
 * Fix beehouses, integration with Better Beehouses.
 * Remove some items.